feat: mark files for firestore storage before storing data in firestore
chore: make sure storing and updating in firestore is working

+++ lib/database/firebase/recipe.dart
@@ -1,5 +1,7 @@
 import 'package:cloud_firestore/cloud_firestore.dart';
+import 'package:image_picker/image_picker.dart';
 import 'package:pambe_ac_ifa/database/cache/cache_client.dart';
+import 'package:pambe_ac_ifa/database/firebase/recipe_images.dart';
 import 'package:pambe_ac_ifa/database/firebase/user.dart';
 import 'package:pambe_ac_ifa/database/interfaces/errors.dart';
 import 'package:pambe_ac_ifa/database/interfaces/firebase.dart';
@@ -35,9 +37,11 @@ class FirebaseRecipeManager
   static const String collectionPath = "recipes";
   FirebaseFirestore db;
   FirebaseUserManager userManager;
+  RemoteRecipeImageManager imageManager;
   CacheClient<RecipeModel> cache;
   CacheClient<PaginatedQueryResult<RecipeLiteModel>> queryCache;
-  FirebaseRecipeManager(this.db, {required this.userManager})
+  FirebaseRecipeManager(this.db,
+      {required this.userManager, required this.imageManager})
       : cache = CacheClient(),
         queryCache = CacheClient(
             cleanupInterval: const Duration(minutes: 1, seconds: 30),
@@ -54,16 +58,24 @@ class FirebaseRecipeManager
     if (cache.has(id)) {
       return Future.value(cache.get(id));
     }
-    final (:data, snapshot: _) = await processDocumentSnapshot(
-        () => db.collection(collectionPath).doc(id).get(),
-        transform: (data, snapshot) async => RecipeModel.fromJson({
-              ...data,
-              "id": snapshot.id,
-              "user":
-                  await userManager.get(data[RecipeFirestoreKeys.userId.name])
-            }));
-    cache.put(id, data);
-    return data;
+    try {
+      final (:data, snapshot: _) = await processDocumentSnapshot(
+          () => db.collection(collectionPath).doc(id).get(),
+          transform: (data, snapshot) async => RecipeModel.fromJson({
+                ...data,
+                "id": snapshot.id,
+                "user":
+                    await userManager.get(data[RecipeFirestoreKeys.userId.name])
+              }));
+      cache.put(id, data);
+      return data;
+    } on ApiError catch (e) {
+      if (e.type == ApiErrorType.resourceNotFound) {
+        return null;
+      } else {
+        rethrow;
+      }
+    }
   }
 
   Future<PaginatedQueryResult<RecipeLiteModel>> getRegularRecipes(
@@ -129,7 +141,14 @@ class FirebaseRecipeManager
       {required String userId}) async {
     final json = recipe.toJson();
     json.remove("user");
+    json.remove("id");
     json[RecipeFirestoreKeys.userId.name] = userId;
+
+    final reserved = imageManager.markRecipeImagesForStorage(
+        current: recipe,
+        userId: userId,
+        former: recipe.remoteId == null ? null : await get(recipe.remoteId!));
+
     String id;
     try {
       if (recipe.remoteId != null) {
@@ -142,6 +161,12 @@ class FirebaseRecipeManager
     } catch (e) {
       throw ApiError(ApiErrorType.storeFailure, inner: e);
     }
+    try {
+      await imageManager.imageManager.process(reserved, userId: userId);
+    } catch (e) {
+      throw ApiError(ApiErrorType.imageManagementFailure, inner: e);
+    }
+
     queryCache.clear();
     cache.markStale(key: id);
     return (await get(id))!;
@@ -149,6 +174,10 @@ class FirebaseRecipeManager
 
   @override
   Future<void> remove(String id) async {
+    final prev = await get(id);
+    if (prev == null) {
+      return;
+    }
     try {
       await db.collection(collectionPath).doc(id).delete();
       queryCache.clear();
